@@ -23,7 +23,7 @@ def read(*rnames):
 
 
 setup(name='dataflake.fakeldap',
-      version='3.1.dev0',
+      version='4.0.dev0',
       description='Mocked-up LDAP connection library',
       long_description=read('README.rst'),
       classifiers=[
@@ -32,13 +32,11 @@ setup(name='dataflake.fakeldap',
         "License :: OSI Approved :: Zope Public License",
         "Operating System :: OS Independent",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.5",
-        "Programming Language :: Python :: 3.6",
         "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
         "Programming Language :: Python :: 3.9",
         "Programming Language :: Python :: 3.10",
-        "Programming Language :: Python :: 3 :: Only",
+        "Programming Language :: Python :: 3.11",
         "Programming Language :: Python :: Implementation :: CPython",
         "Programming Language :: Python :: Implementation :: PyPy",
         "Topic :: Software Development :: Libraries :: Python Modules",
@@ -59,7 +57,7 @@ setup(name='dataflake.fakeldap',
       packages=find_packages('src'),
       package_dir={'': 'src'},
       namespace_packages=['dataflake'],
-      python_requires='>=3.5',
+      python_requires='>=3.7',
       include_package_data=True,
       zip_safe=False,
       install_requires=[
